Finalized the Project Creation through Maven Archetypes.
For now I could not prevent IntelliJ on showing the Sling Module Builder for Modules and so I did show a statement and added a check to make sure that it only works for new Projects (maybe later we add an Archetype that supports modules).
Added a check to make sure an Archetype is selected before allowing the user to proceed to the next step.