All changes for Release 0.6.4.9
Dropped the SNAPSHOT from the version.
Started to work on a better Module handling to unify all the various modules (IntelliJ Project Modules, Maven Project and Server Configuration Module).
Added to support for an override of the Bundle Symbolic Name as Felix is renaming them sometimes and porting all OSGi configuration for non-Maven OSGi bundles to the Facet to work with any OSGi configurations and not just Osmorc.
Made the Server Configuration Verification Action available for any selected Server Configuration and let it run in the Foreground to allow for user interactions.
Moving most of the static text into the Resource Bundle including Debug messages.
Fixed an issue with the Sling Module Builder which was renamed.
Fixed an issue where in the latest IntelliJ the Plugin Settings will not show in the settings.
Fixed an issue with the Debug Mode as it cannot run in the background.
Updated the Documentation to this release and added new topics like Sling Module Facets etc.

Signed-off-by: Andreas Schaefer <hea76404@adobe.com>