Fixed a temporary failure (NPE) due to an issue with IntelliJ not being able to find the IML file
When a project is importent for a short period of time the creation of a Server Connection and the Varification will fail with an NPE because IntelliJ is not able to find the IML file thorugh the Virtual File System.

Using the file path (string), find the parent and use that to obtain the folder.

Signed-off-by: Andreas Schaefer <as@headwire.com>